requirements.txt
boto3
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
